🎭 Dependency update + jest
uvu testing was pretty cool and fast but ran into issues. Move to Jest, much slower :( but works. Resolved dependencies for hours.
