+++ .eslintrc.json
@@ -1,0 +1,10 @@
+ {
+    "root":true,
+    "parser":"@typescript-eslint/parser",
+    "plugins":["@typescript-eslint"],
+    "extends":[
+        "eslint:recommended",
+        "plugin:@typescript-eslint/recommended",
+        "plugin:prettier/recommended"
+    ]
+}
