--- .prettierrc.cjs
@@ -1,7 +1,0 @@
-module.exports = {
-    semi: false,
-    trailingComma: "all",
-    singleQuote: true,
-    printWidth: 120,
-    tabWidth: 4
-};
