--- examples/todomvc/.prettierignore
@@ -1,4 +1,0 @@
-.svelte/**
-static/**
-build/**
-node_modules/**
