--- examples/todomvc/src/global.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="@sveltejs/kit" />
-/// <reference types="svelte" />
-/// <reference types="vite/client" />
