+++ package.json
@@ -25,11 +25,11 @@
         "build-declarations": "tsc -d && node scripts/move-type-declarations.js",
         "build": "npm run build-declarations && rollup -c",
         "dev": "rollup -cw",
-        "test:uvu": "uvu -r ts-node/register tests/tests-uvu",
-        "test:jest": "jest tests/tests-jest",
-        "test:jest:node": "jest tests/tests-jest/node-tests",
-        "test:jest:dom": "jest tests/tests-jest/dom-tests",
-        "test": "npm run test:uvu && npm run test:jest",
+        "test:store": "jest tests/store",
+        "test:persistance": "jest tests/persistance",
+        "test:persistance:dom": "jest tests/tests-jest/dom-tests",
+        "test:persistance:node": "jest tests/tests-jest/node-tests",
+        "test": "jest tests",
         "test:coverage": "c8 --include=src npm test",
         "test:report": "c8 report --reporter=text-lcov > coverage.lcov",
         "prepublishOnly": "npm run build"
@@ -64,7 +64,7 @@
         "eslint": "^8.47.0",
         "eslint-config-prettier": "^9.0.0",
         "eslint-plugin-prettier": "^5.0.0",
-        "eslint-plugin-svelte3": "^4.0.0",
+        "eslint-plugin-svelte": "^2.33.0",
         "gh-pages": "^6.0.0",
         "glob": "^10.3.3",
         "jest": "^29.6.4",
@@ -73,11 +73,11 @@
         "rollup": "^2.46.0",
         "rollup-plugin-sucrase": "^2.1.0",
         "sander": "^0.6.0",
-        "svelte": "^3.37.0",
+        "svelte": "^3.2.0",
+        "svelte-jester": "^3.0.0",
         "ts-jest": "^29.1.1",
         "ts-node": "^9.1.1",
         "typedoc": "^0.24.8",
-        "typescript": "^5.1.6",
-        "uvu": "^0.5.6"
+        "typescript": "^5.1.6"
     }
 }
