+++ src/entity-store.ts
@@ -1,18 +1,16 @@
-import type { Readable, Subscriber, Unsubscriber } from 'svelte/store'
-import { Updater, derived, writable } from 'svelte/store'
-import { getActiveEntity } from './internal/get-active-entity'
-import { getActiveEntityId } from './internal/get-active-entity-id'
-import { getEntities } from './internal/get-entities'
-import { hydrateEntities } from './internal/hydrate-entities'
-import type { Normalized } from './internal/normalize'
-import { normalize } from './internal/normalize'
-import { persistStore } from './internal/persist-store'
-import { removeEntities } from './internal/remove-entities'
-import { setActiveEntity } from './internal/set-active-entity'
-import { setEntities } from './internal/set-entities'
-import { updateEntities } from './internal/update-entities'
-import type { EntityStoreOptions, GetID, ID, Predicate } from './shared'
-import { hydrateActiveId } from './internal/hydrate-active-id'
+import { Readable, Subscriber, Unsubscriber, Updater, derived, writable } from "svelte/store"
+import { EntityStoreOptions, GetID, ID, Predicate } from "./shared"
+import { Normalized, normalize } from "./internal/normalize"
+import { removeEntities } from "./internal/remove-entities"
+import { setEntities } from "./internal/set-entities"
+import { updateEntities } from "./internal/update-entities"
+import { setActiveEntity } from "./internal/set-active-entity"
+import { getEntities } from "./internal/get-entities"
+import { getActiveEntity } from "./internal/get-active-entity"
+import { getActiveEntityId } from "./internal/get-active-entity-id"
+import { hydrateEntities } from "./internal/hydrate-entities"
+import { hydrateActiveId } from "./internal/hydrate-active-id"
+import { persistStore } from "./internal/persist-store"
 
 declare type Invalidator<T> = (value?: T) => void
 declare type Subscribe<T> = (this: void, run: Subscriber<T>, invalidate?: Invalidator<T>) => Unsubscriber
@@ -322,7 +320,7 @@ export function entityStore<T>(
         }
         return createEntityStore<T>(getID, initial)
     } catch (e) {
-        console.error(e)
+        // console.error(e)
         return
     }
 }
