+++ tests/store/index.test.ts
@@ -1,0 +1,5 @@
+import { entityStore } from '../../src'
+
+test('exports the entityStore constructor', () => {
+    expect(entityStore).toBeInstanceOf(Function)
+})
