--- tsconfig.jest.json
@@ -1,16 +1,0 @@
-{
-    "compilerOptions": {
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "noEmitOnError": true,
-        "diagnostics": true,
-        "target": "esnext",
-        "module": "esnext",
-        "declaration": true,
-        "moduleResolution": "node",
-        "forceConsistentCasingInFileNames": true,
-        "skipDefaultLibCheck": true
-    },
-    "include": ["**/*.test.ts"],
-    "exclude": ["node_modules"]
-}
