+++ tsconfig.json
@@ -1,11 +1,4 @@
 {
-    "ts-node": {
-        "transpileOnly": true,
-        "compilerOptions": {
-            "module": "commonjs"
-        },
-        "include": ["./tests/**/*"]
-    },
     "compilerOptions": {
         "noImplicitAny": true,
         "noImplicitThis": true,
